+++ API/elementlist.js
@@ -1,3 +1,3 @@
 
 var ApiGen = ApiGen || {};
-ApiGen.elements = [["c","DatabaseBackup\\Shell\\BackupShell"],["c","DatabaseBackup\\Utility\\BackupExport"],["c","DatabaseBackup\\Utility\\BackupManager"]];
+ApiGen.elements = [["c","DatabaseBackup\\Shell\\BackupShell"],["c","DatabaseBackup\\Utility\\Backup"],["c","DatabaseBackup\\Utility\\BackupExport"],["c","DatabaseBackup\\Utility\\BackupImport"]];