+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # 1.x branch
+## 1.1 branch
+### 1.1.0
+* added the `BackupImport` class;
+* added the `import` shell command;
+* by default the plugin uses the `APP/backups` directory to save the backups files.
+
 ## 1.0 branch
 ### 1.0.3
 * renamed constant for the backup directory.